Add (slightly buggy) video export
Videos can now be generated by pressing "Create video".
At the moment, this only works in chrome, due to limitations
in CCapture.js, but it should be possible to generate image
sequences and gifs in other browsers.

For some reason, you have to press "Next Config" for the movie
to download at the end...