Fix video save issue and add video options
Now possible to save trajectory as either webm video, gif, or as
a collection of images in either png or jpg format. Saving webm
still works only on Chrome. Saving as gif uses a webworker, so
it will not work unless done on a webserver (simply running
"python3 -m http.server" worked for me in order to test locally)